+++ Source/ASPagerNode.m
@@ -16,6 +16,10 @@
 //
 
 #import <AsyncDisplayKit/ASPagerNode.h>
+#import <AsyncDisplayKit/ASPagerNode+Beta.h>
+
+#import <AsyncDisplayKit/ASCollectionGalleryLayoutDelegate.h>
+#import <AsyncDisplayKit/ASCollectionNode+Beta.h>
 #import <AsyncDisplayKit/ASDelegateProxy.h>
 #import <AsyncDisplayKit/ASDisplayNode+FrameworkPrivate.h>
 #import <AsyncDisplayKit/ASDisplayNode+Subclasses.h>
@@ -25,10 +29,8 @@
 #import <AsyncDisplayKit/ASCollectionView+Undeprecated.h>
 #import <AsyncDisplayKit/UIResponder+AsyncDisplayKit.h>
 
-@interface ASPagerNode () <ASCollectionDataSource, ASCollectionDelegate, ASCollectionDelegateFlowLayout, ASDelegateProxyInterceptor>
+@interface ASPagerNode () <ASCollectionDataSource, ASCollectionDelegate, ASCollectionDelegateFlowLayout, ASDelegateProxyInterceptor, ASCollectionGalleryLayoutSizeProviding>
 {
-  ASPagerFlowLayout *_flowLayout;
-
   __weak id <ASPagerDataSource> _pagerDataSource;
   ASPagerNodeProxy *_proxyDataSource;
   struct {
@@ -65,8 +67,15 @@
 {
   ASDisplayNodeAssert([flowLayout isKindOfClass:[ASPagerFlowLayout class]], @"ASPagerNode requires a flow layout.");
   self = [super initWithCollectionViewLayout:flowLayout];
-  if (self != nil) {
-    _flowLayout = flowLayout;
+  return self;
+}
+
+- (instancetype)initUsingAsyncCollectionLayout
+{
+  ASCollectionGalleryLayoutDelegate *layoutDelegate = [[ASCollectionGalleryLayoutDelegate alloc] initWithScrollableDirections:ASScrollDirectionHorizontalDirections];
+  self = [super initWithLayoutDelegate:layoutDelegate layoutFacilitator:nil];
+  if (self) {
+    layoutDelegate.sizeProvider = self;
   }
   return self;
 }
@@ -126,6 +135,14 @@
     return NSNotFound;
   }
   return indexPath.row;
+}
+
+#pragma mark - ASCollectionGalleryLayoutSizeProviding
+
+- (CGSize)sizeForElements:(ASElementMap *)elements
+{
+  ASDisplayNodeAssertMainThread();
+  return self.bounds.size;
 }
 
 #pragma mark - ASCollectionDataSource