+++ Source/Details/ASCollectionLayoutContext.m
@@ -37,7 +37,6 @@
 {
   self = [super init];
   if (self) {
-    ASDisplayNodeAssertTrue([layoutDelegateClass conformsToProtocol:@protocol(ASCollectionLayoutDelegate)]);
     _viewportSize = viewportSize;
     _scrollableDirections = scrollableDirections;
     _elements = elements;